requirements.txt
Django~=4.1.7
django-environ~=0.10.0
pymongo~=4.3.3
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.2.2
drf-spectacular[sidecar]~=0.26.1
django-cors-headers~=3.14.0
